Load metadata from directory things
Sometimes we're downloading older stuff from thingiverse that doesn't
have a parseable LICENSE.txt.

Here we add a metadata file called =miniverse.json= where you can
specify the id, title, author, and license which overrides what we would
parse from the ~LICENSE.txt~.

This means we have a way of storing
things even when the metadata parse would otherwise have failed and
called them unthings.

Currently this is factored a little oddly - the json is turned into a
license string that can be regexed; instead it should probably be more
like the metadata loader loads from json or parses the license, but this
gets us going.